pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nova_demo"
authors = [
    {name = "Magnus Kühne", email = "magnus.kuehne@serapion.net"},
]
description = "NOVA Demo Server"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
]
dynamic = ["version"]
dependencies = [
    "openai>=1.6.1,<2.0.0",
    "fastapi>=0.108.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "uvicorn>=0.23.2,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "fastapi-auth0>=0.5.0,<1.0.0",
    "aiohttp>=3.8.5,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-dotenv>=0.5.2,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "mongomock_motor>=0.0.21,<1.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "flake8>=7.0.0,<8.0.0"
]

ci = [
    "pytest>=7.4.2,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-dotenv>=0.5.2,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "mongomock_motor>=0.0.21,<1.0.0",
    "python-semantic-release>=9.0.0,<10.0.0",
    "flake8>=7.0.0,<8.0.0"
]

[tool.setuptools.dynamic]
version = {attr = "nova_demo.__version__"}

[tool.pytest.ini_options]
asyncio_mode = "strict"
env_files = [
    ".env.pytest"
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.black]
line-length = 120

[tool.semantic_release]
version_variable = "nova_demo/__init__.py:__version__"
